# openapi/api-v1.yaml
openapi: 3.1.0
info:
  title: ILM Red API
  description: |
    Cloud-native API for digital knowledge management platform.

    ## Authentication

    The API supports two authentication methods:

    1. **Bearer Token (JWT)**: Include the JWT in the `Authorization` header
       ```
       Authorization: Bearer eyJhbGciOiJSUzI1NiI...
       ```

    2. **API Key**: Include your API key in the `X-API-Key` header
       ```
       X-API-Key: ilm_live_abc123...
       ```

    ## Rate Limiting

    | Tier       | Requests/min | AI Tokens/day |
    |------------|--------------|---------------|
    | Free       | 60           | 10,000        |
    | Premium    | 300          | 100,000       |
    | Enterprise | 1,000        | Unlimited     |

    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`
    - `X-RateLimit-Remaining`
    - `X-RateLimit-Reset`

    ## Pagination

    List endpoints support pagination with `page` and `limit` query parameters.
    Response includes a `pagination` object with total count and page info.

  version: 1.0.0
  contact:
    name: ILM Red Support
    email: support@ilm-red.com
    url: https://docs.ilm-red.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.ilm-red.com/v1
    description: Production
  - url: https://api-staging.ilm-red.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Development
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /books:
    get:
      tags:
        - Books
      summary: List books
      operationId: listBooks
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: visibility
          in: query
          schema:
            type: string
            enum: [public, private, friends]
        - name: category
          in: query
          schema:
            type: array
            items:
              $ref: '#/components/schemas/BookCategory'
        - name: language
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [createdAt, title, rating, views]
            default: createdAt
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of books
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Book'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Books
      summary: Upload a new book
      operationId: createBook
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - title
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF, EPUB, or TXT file (max 500MB)
                title:
                  type: string
                  maxLength: 500
                author:
                  type: string
                  maxLength: 200
                description:
                  type: string
                  maxLength: 5000
                category:
                  $ref: '#/components/schemas/BookCategory'
                visibility:
                  type: string
                  enum: [public, private, friends]
                  default: private
                language:
                  type: string
                  description: ISO 639-1 language code
      responses:
        '201':
          description: Book created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          $ref: '#/components/responses/ValidationError'
        '413':
          description: File too large
  /books/{bookId}:
    get:
      tags:
        - Books
      summary: Get book details
      operationId: getBook
      parameters:
        - $ref: '#/components/parameters/BookIdParam'
      responses:
        '200':
          description: Book details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Books
      summary: Update book metadata
      operationId: updateBook
      parameters:
        - $ref: '#/components/parameters/BookIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 500
                author:
                  type: string
                  maxLength: 200
                description:
                  type: string
                  maxLength: 5000
                category:
                  $ref: '#/components/schemas/BookCategory'
                visibility:
                  type: string
                  enum: [public, private, friends]
      responses:
        '200':
          description: Book updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Books
      summary: Delete a book
      operationId: deleteBook
      parameters:
        - $ref: '#/components/parameters/BookIdParam'
      responses:
        '204':
          description: Book deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /users/{userId}/books:
    get:
      tags:
        - Users
      summary: Get user's public books
      operationId: getUserBooks
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: User's books
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Book'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

  # ============= SEARCH =============
...
  /search/books:
    get:
      tags:
        - Search
      summary: Search books
      operationId: searchBooks
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: array
            items:
              $ref: '#/components/schemas/BookCategory'
        - name: language
          in: query
          schema:
            type: string
        - name: author
          in: query
          schema:
            type: string
        - name: minRating
          in: query
          schema:
            type: number
            minimum: 1
            maximum: 5
        - name: sort
          in: query
          schema:
            type: string
            enum: [relevance, title, date, rating, views]
            default: relevance
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Book search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BookSearchResult'
                  facets:
                    $ref: '#/components/schemas/SearchFacets'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
  /clubs:
    get:
      tags:
        - Clubs
      summary: Discover book clubs
      operationId: listClubs
      parameters:
        - name: visibility
          in: query
          schema:
            type: string
            enum: [public]
        - name: q
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of clubs
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Club'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Clubs
      summary: Create book club
      operationId: createClub
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  maxLength: 100
                description:
                  type: string
                  maxLength: 1000
                visibility:
                  type: string
                  enum: [public, private]
                  default: public
                joinPolicy:
                  type: string
                  enum: [open, approval, invite_only]
                  default: open
      responses:
        '201':
          description: Club created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Club'
  /clubs/{clubId}:
    get:
      tags:
        - Clubs
      summary: Get club details
      operationId: getClub
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Club details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClubDetail'

    patch:
      tags:
        - Clubs
      summary: Update club
      operationId: updateClub
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                visibility:
                  type: string
                  enum: [public, private]
                joinPolicy:
                  type: string
                  enum: [open, approval, invite_only]
      responses:
        '200':
          description: Club updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Club'

    delete:
      tags:
        - Clubs
      summary: Delete club
      operationId: deleteClub
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Club deleted
  /clubs/{clubId}/members:
    get:
      tags:
        - Clubs
      summary: List club members
      operationId: listClubMembers
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Club members
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ClubMember'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Clubs
      summary: Join club
      operationId: joinClub
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Joined club
        '202':
          description: Join request submitted (pending approval)
  /clubs/{clubId}/members/{userId}:
    patch:
      tags:
        - Clubs
      summary: Update member role
      operationId: updateClubMember
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
              properties:
                role:
                  type: string
                  enum: [admin, moderator, member]
      responses:
        '200':
          description: Member role updated

    delete:
      tags:
        - Clubs
      summary: Remove member
      operationId: removeClubMember
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Member removed
  /clubs/{clubId}/invites:
    post:
      tags:
        - Clubs
      summary: Invite to club
      operationId: inviteToClub
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                role:
                  type: string
                  enum: [admin, moderator, member]
                  default: member
                message:
                  type: string
                  maxLength: 500
      responses:
        '201':
          description: Invite sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClubInvite'
  /clubs/{clubId}/books:
    get:
      tags:
        - Clubs
      summary: List club books
      operationId: listClubBooks
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Club books
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Book'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Clubs
      summary: Add book to club
      operationId: addBookToClub
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - bookId
              properties:
                bookId:
                  type: string
      responses:
        '201':
          description: Book added

  # ============= PROGRESS =============
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    BookIdParam:
      name: bookId
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
            timestamp:
              type: string
              format: date-time
            documentation:
              type: string
              format: uri
      example:
        error:
          code: "BOOK_NOT_FOUND"
          message: "The requested book does not exist or you don't have access"
          requestId: "req_abc123xyz789"
          timestamp: "2025-07-15T10:30:00Z"
          documentation: "https://api.ilm-red.com/docs/errors#BOOK_NOT_FOUND"
    ValidationError:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - details
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasMore:
          type: boolean
      example:
        page: 1
        limit: 20
        total: 100
        totalPages: 5
        hasMore: true
...
    BookCategory:
      type: string
      enum:
        - religion
        - islamic_studies
        - quran_studies
        - hadith
        - fiqh
        - history
        - biography
        - philosophy
        - science
        - technology
        - education
        - self_help
        - fiction
        - children
        - languages
        - other
    Book:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        author:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        category:
          $ref: '#/components/schemas/BookCategory'
        language:
          type: string
        visibility:
          type: string
          enum: [public, private, friends, club]
        thumbnailUrl:
          type: string
          format: uri
          nullable: true
        stats:
          type: object
          properties:
            viewCount:
              type: integer
            favoriteCount:
              type: integer
            rating:
              type: number
              nullable: true
        status:
          type: string
          enum: [processing, active, deleted]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        id: "ffe66482-4ee7-4de8-8c94-38d5414c1d17"
        title: "Charlie and the Chocolate Factory"
        author: "Roald Dahl"
        description: "Charlie and the Chocolate Factory is a classic children's novel by Roald Dahl, first published in 1964."
        category: "children"
        language: "en"
        visibility: "public"
        thumbnailUrl: "https://cdn.ilm-red.com/thumbnails/charlie-chocolate.png"
        stats:
          viewCount: 2
          favoriteCount: 5
          rating: 4.5
        status: "active"
        createdAt: "2025-07-10T18:00:29.753Z"
        updatedAt: "2025-07-13T20:22:42.779Z"
    BookDetail:
      allOf:
        - $ref: '#/components/schemas/Book'
        - type: object
          properties:
            owner:
              $ref: '#/components/schemas/PublicUserProfile'
            file:
              type: object
              properties:
                size:
                  type: integer
                mimeType:
                  type: string
                pageCount:
                  type: integer
                  nullable: true
            aiEnabled:
              type: boolean
            isFavorite:
              type: boolean
            userRating:
              type: integer
              nullable: true
    BookSearchResult:
      allOf:
        - $ref: '#/components/schemas/Book'
        - type: object
          properties:
            score:
              type: number
            highlights:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
...
    PublicUserProfile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          format: uri
          nullable: true
        bio:
          type: string
          nullable: true
      example:
        id: "61e3db52-fac1-4c00-8fbf-75287ed8084a"
        username: "reader_61e3db52"
        displayName: "Alex Smith"
        avatarUrl: null
        bio: "Book enthusiast and avid reader"
...
    SearchFacets:
      type: object
      properties:
        categories:
          type: array
          items:
            type: object
            properties:
              value:
                type: string
              count:
                type: integer
        languages:
          type: array
          items:
            type: object
            properties:
              value:
                type: string
              count:
                type: integer
        authors:
          type: array
          items:
            type: object
            properties:
              value:
                type: string
              count:
                type: integer
...
    Club:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          format: uri
          nullable: true
        visibility:
          type: string
          enum: [public, private]
        joinPolicy:
          type: string
          enum: [open, approval, invite_only]
        memberCount:
          type: integer
        bookCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    ClubDetail:
      allOf:
        - $ref: '#/components/schemas/Club'
        - type: object
          properties:
            owner:
              $ref: '#/components/schemas/PublicUserProfile'
            isMember:
              type: boolean
            memberRole:
              type: string
              nullable: true
    ClubMember:
      type: object
      properties:
        userId:
          type: string
        role:
          type: string
          enum: [owner, admin, moderator, member]
        joinedAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/PublicUserProfile'
    ClubInvite:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
        status:
          type: string
          enum: [pending, accepted, expired]
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...